Jump the gun
I briefly forgot the life-saving Tic Tac Toe project article that
discussed how to build software logic. It specifically exhorts
beginners to write the logic before the DOM. I had every intention of
doing so but wanted help visualizing the moving pieces, so I build the
core of the DOM. But after the default elements of the DOM were built,
I promptly forgot this good advice and tried to build the logic from
an event listener.

The code submitted is not long and can be ignored safely. It might be
useful after the logic is written, but for now it will sleep.