Add working logic
The todo app works fine in the console when the script is added directly
to index.html. Through the console, it is possible to add and delete
projects. Likewise, it is possible to add, delete, rename, reschedule,
and complete tasks.

The current solution does not work once exported and imported. Perhaps
my lackluster knowledge of Webpack is to blame. I will try to connect
the UI to the logic that this commit is bringing.