Fix a bug related to hardcode
To test the app, it was easier to hardcode some values. But
after implementing the localStorage functionality, a bugish behavior
inserted itself.

In short, a project was created (and therefore selected), and a
task was added. But upon reloading the page, the project could not be
created by the hardcode (it was being read from localStorage). Then,
a task was still added, despite the correct project not being
selected.

I fixed this non-bug by selecting the project I was attempting to
create. That way, on reload, the project is not created anew, but
it is still selected before the relevant task is added.